WIP: refactor: moved the file_format_checker
